Created BlogPostForm componenet to pass reuse form and pass state to edit and create screens. updated edit and create screens and removed now unused code
