+++ src/screens/CreateScreen.js
@@ -1,52 +1,14 @@
-import React, { useContext, useState } from 'react';
-import { Text, TextInput, View, StyleSheet, Button } from 'react-native';
+import React, { useContext } from 'react';
+import { StyleSheet } from 'react-native';
 import { Context } from '../context/BlogContext';
+import BlogPostForm from '../components/BlogPostForm';
 
 const CreateScreen = ({ navigation }) => {
-  const [title, setTitle] = useState('');
-  const [content, setContent] = useState('');
   const { addBlogPost } = useContext(Context);
 
-  return (
-    <View>
-      <Text style={styles.label}>Enter Title:</Text>
-      <TextInput
-        style={styles.input}
-        value={title}
-        onChangeText={(text) => setTitle(text)}
-      />
-      <Text style={styles.label}>Enter Content:</Text>
-      <TextInput
-        style={styles.input}
-        value={content}
-        onChangeText={(text) => setContent(text)}
-      />
-      <Button
-        title="Add a Post"
-        onPress={() => {
-          addBlogPost(title, content, () => {
-            navigation.navigate('Index');
-          });
-        }}
-      />
-    </View>
-  );
+  return <BlogPostForm />;
 };
 
-const styles = StyleSheet.create({
-  input: {
-    fontSize: 18,
-    borderWidth: 1,
-    borderColor: 'black',
-    marginBottom: 15,
-    padding: 5,
-    margin: 5
-  },
-  label: {
-    fontSize: 20,
-    marginBottom: 5,
-    marginLeft: 5
-  }
-});
+const styles = StyleSheet.create({});
 
 export default CreateScreen;
