+++ src/screens/EditScreen.js
@@ -1,6 +1,7 @@
-import React, { useState, useContext } from 'react';
-import { Text, View, StyleSheet, TextInput } from 'react-native';
+import React, { useContext } from 'react';
+import { StyleSheet } from 'react-native';
 import { Context } from '../context/BlogContext';
+import BlogPostForm from '../components/BlogPostForm';
 
 const EditScreen = ({ navigation }) => {
   const { state } = useContext(Context);
@@ -9,23 +10,7 @@ const EditScreen = ({ navigation }) => {
     (blogPost) => blogPost.id === navigation.getParam('id')
   );
 
-  const [title, setTitle] = useState(blogPost.title);
-  const [content, setContent] = useState(blogPost.content);
-
-  return (
-    <View>
-      <Text>Edit Title:</Text>
-      <TextInput
-        value={title}
-        onChangeText={(newTitle) => setTitle(newTitle)}
-      />
-      <Text>Edit Content:</Text>
-      <TextInput
-        value={content}
-        onChangeText={(newContent) => setContent(newContent)}
-      />
-    </View>
-  );
+  return <BlogPostForm />;
 };
 
 const styles = StyleSheet.create({});
